Add Higher Order Interpolation explainer
After a few minutes of sketching on the back of an envelope and a week
of evenings translating these sketches into a coherent explanation, this
outlines how not to do HOI, one way to do HOI, and provides an example.
This could be greatly improved with a bunch of fun interactive
animations, but those will have to come later.